make Zsh-compatible without breaking Bash
Zsh defines $path as an array containing the contents of $PATH, for
easier manipulation without having to splice on colons (and similar for
other ALL_CAPS vs lower_case variable pairs). This commit changes that
variable name to $vars_path, changes the shebang from "bash" to "sh",
adds to the ShellCheck disable directive, and explains the extended
directive.