naive ccr->chaser update based on `chaser --help`
I'll probably have to update something later. In particular, it looks
like `chaser install` and `chaser search` are limited to the CCR. Thus I
may have to update `pm` to process its data files as scripts instead of
command lists. But I'm leaving it as-is for now until I actually run
into the possible problems.