trade broken pacman installing for broken CCR
The `chaser install` command only works with CCR packages and `pacman
-S` only works with non-CCR packages. Since CCR-packages are by
definition nonessential for the system, it's better to break installing
them with `pm` instead of breaking non-CCR package install. This also
shows a TODO message with directions for manual non-`pm` install of CCR
packages so the user knows what's going on and can work around this
limitation.

Note that this seems to be the only remaining ccr->chaser problem after
the last two `pm` data updates. I tested all commands (if, in, rm, s,
up) with CCR and non-CCR packages as applicable, and install is the only
one that didn't work.