auto-lock: switch from qdbus(-qt5) to dbus-send
while the former is friendlier for interactive use, the latter is more
likely to be on random *nix boxes. the script should now work
correctly on pretty much any system that has dbus, not just the ones
that use the qt framework. no idea if macos supports any freedesktop
standards like this.